Routes now match swagger spec
