+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
   namespace 'identifiers' do
-    match '/ppon', to: 'organisations#create', via: [:post]
+    namespace 'id' do
+      match '/ppon', to: 'organisations#create', via: [:post]
+    end
   end
   post '*unmatched_route', to: 'application#not_found'
 end
